added execution time and finished Solve methods

+++ Extensions/RangeExtension.cs
@@ -18,7 +18,7 @@ namespace PrimeNumbersThreaded.Extensions
             var counter = start;
             while (counter <= end)
             {
-                yield return (int) counter;
+                yield return counter;
                 counter += step;
             }
         }
